Move process utilities to subspace-process
