+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/cluster/cache.rs
@@ -14,7 +14,7 @@ use std::time::Duration;
 use subspace_farmer::cluster::cache::cache_service;
 use subspace_farmer::cluster::nats_client::NatsClient;
 use subspace_farmer::disk_piece_cache::DiskPieceCache;
-use subspace_networking::utils::AsyncJoinOnDrop;
+use subspace_process::AsyncJoinOnDrop;
 
 /// Interval between cache self-identification broadcast messages
 pub(super) const CACHE_IDENTIFICATION_BROADCAST_INTERVAL: Duration = Duration::from_secs(30);
