+++ crates/subspace-farmer/src/bin/subspace-farmer/commands/farm.rs
@@ -51,9 +51,7 @@ use subspace_farmer_components::reading::ReadSectorRecordChunksMode;
 use subspace_kzg::Kzg;
 use subspace_metrics::{RegistryAdapter, start_prometheus_metrics_server};
 use subspace_networking::utils::piece_provider::PieceProvider;
-use subspace_networking::utils::{
-    AsyncJoinOnDrop, run_future_in_dedicated_thread, shutdown_signal,
-};
+use subspace_process::{AsyncJoinOnDrop, run_future_in_dedicated_thread, shutdown_signal};
 use subspace_proof_of_space::Table;
 use tracing::{Instrument, error, info, info_span, warn};
