+++ crates/subspace-farmer/src/single_disk_farm/plot_cache.rs
@@ -18,8 +18,8 @@ use subspace_core_primitives::sectors::SectorIndex;
 use subspace_farmer_components::file_ext::FileExt;
 use subspace_farmer_components::sector::SectorMetadataChecksummed;
 use subspace_networking::libp2p::kad::RecordKey;
-use subspace_networking::utils::AsyncJoinOnDrop;
 use subspace_networking::utils::multihash::ToMultihash;
+use subspace_process::AsyncJoinOnDrop;
 use thiserror::Error;
 use tokio::task;
 use tracing::{debug, info, warn};
