+++ crates/subspace-networking/examples/metrics.rs
@@ -8,9 +8,8 @@ use prometheus_client::registry::Registry;
 use std::sync::Arc;
 use std::time::Duration;
 use subspace_metrics::{RegistryAdapter, start_prometheus_metrics_server};
-use subspace_networking::utils::shutdown_signal;
 use subspace_networking::{Config, Node};
-use subspace_process::init_logger;
+use subspace_process::{init_logger, shutdown_signal};
 use tokio::time::sleep;
 use tracing::{error, info};
