+++ crates/subspace-node/src/commands/run.rs
@@ -35,8 +35,7 @@ use sp_messenger::messages::ChainId;
 use std::env;
 use std::sync::Arc;
 use subspace_metrics::{RegistryAdapter, start_prometheus_metrics_server};
-use subspace_networking::utils::{raise_fd_limit, run_future_in_dedicated_thread};
-use subspace_process::init_logger;
+use subspace_process::{init_logger, raise_fd_limit, run_future_in_dedicated_thread};
 use subspace_runtime::{Block, RuntimeApi};
 use subspace_service::config::ChainSyncMode;
 use tracing::{debug, error, info, info_span, warn};
